+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/crossmodel/crossmodel-core/compare/v0.10.9...v0.11.0) (2025-11-13)
+
+
+### Features
+
+* Multiple identifiers ([#210](https://github.com/crossmodel/crossmodel-core/issues/210)) ([996e731](https://github.com/crossmodel/crossmodel-core/commit/996e731abcc7aeffd4633ee51d99fcb5fa79c17d))
+
+
+
+
+
 ## [0.10.9](https://github.com/crossmodel/crossmodel-core/compare/v0.10.8...v0.10.9) (2025-11-06)