+++ packages/model-service/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.11.0](https://github.com/crossmodel/crossmodel-core/compare/v0.10.9...v0.11.0) (2025-11-13)
+
+**Note:** Version bump only for package @crossmodel/model-service
+
+
+
+
+
 ## [0.10.9](https://github.com/crossmodel/crossmodel-core/compare/v0.10.8...v0.10.9) (2025-11-06)
 
 **Note:** Version bump only for package @crossmodel/model-service